requirements.txt
# Core AI and ML Framework
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-core==0.1.10
langchain-experimental==0.0.47
langchain-huggingface==0.0.6
langgraph==0.0.20

# OpenAI and Groq Integration
openai==1.3.7

# Streamlit Web Framework
streamlit==1.28.1

# Database
pymongo==4.6.0

# Vector Database and Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
unstructured==0.11.8

# Data Validation and Serialization
pydantic==2.5.0

# Environment and Configuration
python-dotenv==1.0.0

# Text Processing and Utilities
regex==2023.10.3
typing-extensions==4.8.0

# Optional: For enhanced document processing
pdfplumber==0.10.3
python-multipart==0.0.6

# Development and Testing (Optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0 